Move RewriteCtx into a trait.
I'd like to separate the RewriteCtx implemetation from it's contract.
The objective is not to aid testing or to facilitate multiple
implementations of RewriteCtx. The objective is only to make it
easier to read the API contract without the implementations.
